SSL & sessionState
SSL is on, sessionState is enabled.  I coded (I believe) all the back end and I threw in a generic button that can be coded on page load to only display if "flag" (sessionState["Flag"]) if roll = "Admin".  I believe I already have the specific if statements written, just needs the UI front end written to display or not to display.

Whomever has taken over the UI or if Bob is still doing it I believe the it should be fine to simply have them press the back button to go back to the display page but if you think a back button is required slap one on the page and it can be SSL redirect coded to go back to display page.